Configure syntastic to use eslint for javascript
Use locally installed version of eslint and eslint plugins. Otherwise,
eslint has to be installed globally, and any plugins used in a
particular project also have to be installed globally. This also ensures
a consistent version of eslint it used, that works with the eslint
config in the project.

This requires a hacky fix so that Syntastic can determine whether eslint
is available, by setting the `exec` to a valid binary.