Unmap default shift+alt+down to extension.vim_cmd+alt+down key binding
Not sure what this default is intended to do, but it interferes with
multiple cursors.

By unmapping this, shift+alt+down correctly inserts a new cursor on the
following line. With the default in place, shift+alt+down doesn't seem
to do anything at all in either normal or insert mode.